requirements.txt
# ==== A ====
aiofiles==23.2.1
aiohttp==3.9.0
APScheduler==3.10.4
alphabet-detector==0.0.7
asyncio==3.4.3

# ==== B ====
beautifulsoup4==4.12.2
bleach==6.1.0

# ==== C ====
cachetools==5.3.2

# ==== D ====
dateparser==1.2.0

# ==== E ====
emoji==2.8.0
emoji-country-flag==1.3.2

# ==== F ====
ffmpeg-python
future==0.18.3

# ==== G ====
gpytranslate==1.5.1

# ==== H ====
html2text==2020.1.16
httpx[http2]

# ==== J ====
jikanpy==4.3.2

# ==== L ====
lxml==4.9.3

# ==== M ====
markdown2==2.4.10
motor==3.3.2

# ==== N ====
nekos.py==1.1.0

# ==== O ====
opencv-python-headless==4.8.1.78

# ==== P ====
Pillow==9.5.0
psutil==5.9.6
psycopg2-binary==2.9.9
pynewtonmath==1.0.1
pymongo[srv]==4.6.0
pyrate-limiter==3.1.0
pytz==2023.3.post1
python-arq==6.0.5
pyrogram==2.0.106
python-telegram-bot==20.6

# ==== R ====
regex==2023.10.3
requests==2.31.0

# ==== S ====
shortuuid==1.0.11
SQLAlchemy==1.4.49

# ==== T ====
telegraph==2.2.0
telethon==1.32.1
tgcrypto==1.2.5

# ==== U ====
Unidecode==1.3.7
